bootpack.cの分割　分割コンパイルの実装
